Hup 72,74,75 (#244)
* HUP-66 PMS API selected Country null fix

* HUP-72

Deleted product edit button change to deleted button

* HUP-74

The extra language we create when creating the product can be the same as the default locale language if it has the same name and description. If not it can not be the same.

* HUP-75

The locale of the extra language we created while creating the product is not initialized.